+++ elasticsearch_dsl/document.py
@@ -1,5 +1,6 @@
 import re
 
+from elasticsearch.exceptions import NotFoundError, RequestError
 from six import iteritems, add_metaclass
 
 from .field import Field
@@ -44,7 +45,7 @@ class DocTypeOptions(object):
         self._using = getattr(meta, 'using', None)
 
         # get doc_type name, if not defined take the name of the class and
-        # tranform it to lower_case
+        # transform it to lower_case
         doc_type = getattr(meta, 'doc_type',
                 re.sub(r'(.)([A-Z])', r'\1_\2', name).lower())
 
@@ -97,8 +98,11 @@ class DocType(ObjectBase):
     def __init__(self, meta=None, **kwargs):
         meta = meta or {}
         for k in list(kwargs):
-            if k.startswith('_'):
+            if k.startswith('_') and k[1:] in META_FIELDS:
                 meta[k] = kwargs.pop(k)
+
+        if self._doc_type.index:
+            meta.setdefault('_index', self._doc_type.index)
         super(AttrDict, self).__setattr__('meta', ResultMeta(meta))
 
         super(DocType, self).__init__(**kwargs)
@@ -115,6 +119,13 @@ class DocType(ObjectBase):
         if name.startswith('_') and name[1:] in META_FIELDS:
             return getattr(self.meta, name[1:])
         return super(DocType, self).__getattr__(name)
+
+    def __repr__(self):
+        return '%s(%s)' % (
+            self.__class__.__name__,
+            ', '.join('%s=%r' % (key, getattr(self.meta, key)) for key in
+                      ('index', 'doc_type', 'id') if key in self.meta)
+        )
 
     def __setattr__(self, name, value):
         if name.startswith('_') and name[1:] in META_FIELDS:
@@ -147,6 +158,54 @@ class DocType(ObjectBase):
         return cls.from_es(doc)
 
     @classmethod
+    def mget(cls, docs, using=None, index=None, raise_on_error=True,
+             missing='none', **kwargs):
+        if missing not in ('raise', 'skip', 'none'):
+            raise ValueError("'missing' must be 'raise', 'skip', or 'none'.")
+        es = connections.get_connection(using or cls._doc_type.using)
+        body = {'docs': [doc if isinstance(doc, dict) else {'_id': doc}
+                         for doc in docs]}
+        results = es.mget(
+            body,
+            index=index or cls._doc_type.index,
+            doc_type=cls._doc_type.name,
+            **kwargs
+        )
+
+        objs, error_docs, missing_docs = [], [], []
+        for doc in results['docs']:
+            if doc.get('found'):
+                if error_docs or missing_docs:
+                    # We're going to raise an exception anyway, so avoid an
+                    # expensive call to cls.from_es().
+                    continue
+
+                objs.append(cls.from_es(doc))
+
+            elif doc.get('error'):
+                if raise_on_error:
+                    error_docs.append(doc)
+                if missing == 'none':
+                    objs.append(None)
+
+            # The doc didn't cause an error, but the doc also wasn't found.
+            elif missing == 'raise':
+                missing_docs.append(doc)
+            elif missing == 'none':
+                objs.append(None)
+
+        if error_docs:
+            error_ids = [doc['_id'] for doc in error_docs]
+            message = 'Required routing/parent not provided for documents %s.'
+            message %= ', '.join(error_ids)
+            raise RequestError(400, message, error_docs)
+        if missing_docs:
+            missing_ids = [doc['_id'] for doc in missing_docs]
+            message = 'Documents %s not found.' % ', '.join(missing_ids)
+            raise NotFoundError(404, message, missing_docs)
+        return objs
+
+    @classmethod
     def from_es(cls, hit):
         # don't modify in place
         meta = hit.copy()
@@ -156,7 +215,7 @@ class DocType(ObjectBase):
             for k, v in iteritems(meta.pop('fields')):
                 if k == '_source':
                     doc.update(v)
-                if k.startswith('_'):
+                if k.startswith('_') and k[1:] in META_FIELDS:
                     meta[k] = v
                 else:
                     doc[k] = v
@@ -191,23 +250,24 @@ class DocType(ObjectBase):
 
     def to_dict(self, include_meta=False):
         d = super(DocType, self).to_dict()
-        if include_meta:
-            meta = dict(
-                ('_' + k, self.meta[k])
-                for k in DOC_META_FIELDS
-                if k in self.meta
-            )
+        if not include_meta:
+            return d
 
-            # in case of to_dict include the index unlike save/update/delete
-            if 'index' in self.meta:
-                meta['_index'] = self.meta.index
-            elif self._doc_type.index:
-                meta['_index'] = self._doc_type.index
+        meta = dict(
+            ('_' + k, self.meta[k])
+            for k in DOC_META_FIELDS
+            if k in self.meta
+        )
 
-            meta['_type'] = self._doc_type.name
-            meta['_source'] = d
-            d = meta
-        return d
+        # in case of to_dict include the index unlike save/update/delete
+        if 'index' in self.meta:
+            meta['_index'] = self.meta.index
+        elif self._doc_type.index:
+            meta['_index'] = self._doc_type.index
+
+        meta['_type'] = self._doc_type.name
+        meta['_source'] = d
+        return meta
 
     def update(self, using=None, index=None, **fields):
         es = self._get_connection(using)