+++ elasticsearch_dsl/utils.py
@@ -26,7 +26,7 @@ def _make_dsl_class(base, name, params_def=None, suffix=''):
 
 class AttrList(object):
     def __init__(self, l, obj_wrapper=None):
-        # make iteables into lists
+        # make iterables into lists
         if not isinstance(l, list):
             l = list(l)
         self._l_ = l
@@ -40,6 +40,9 @@ class AttrList(object):
             return other._l_ == self._l_
         # make sure we still equal to a dict with the same data
         return other == self._l_
+
+    def __ne__(self, other):
+        return not self == other
 
     def __getitem__(self, k):
         l = self._l_[k]
@@ -62,6 +65,12 @@ class AttrList(object):
 
     def __getattr__(self, name):
         return getattr(self._l_, name)
+
+    def __getstate__(self):
+        return (self._l_, self._obj_wrapper)
+
+    def __setstate__(self, state):
+        self._l_, self._obj_wrapper = state
 
 
 class AttrDict(object):
@@ -91,11 +100,20 @@ class AttrDict(object):
         # make sure we still equal to a dict with the same data
         return other == self._d_
 
+    def __ne__(self, other):
+        return not self == other
+
     def __repr__(self):
         r = repr(self._d_)
         if len(r) > 60:
             r = r[:60] + '...}'
         return r
+
+    def __getstate__(self):
+        return (self._d_, )
+
+    def __setstate__(self, state):
+        super(AttrDict, self).__setattr__('_d_', state[0])
 
     def __getattr__(self, attr_name):
         try:
@@ -182,7 +200,7 @@ class DslBase(object):
         - to_dict method to serialize into dict (to be sent via elasticsearch-py)
         - basic logical operators (&, | and ~) using a Bool(Filter|Query) TODO:
           move into a class specific for Query/Filter
-        - respects the definiton of the class and (de)serializes it's
+        - respects the definition of the class and (de)serializes it's
           attributes based on the `_param_defs` definition (for example turning
           all values in the `must` attribute into Query objects)
     """
@@ -195,10 +213,10 @@ class DslBase(object):
         except KeyError:
             raise UnknownDslObject('DSL class `%s` does not exist in %s.' % (name, cls._type_name))
 
-    def __init__(self, **params):
+    def __init__(self, _expand__to_dot=True, **params):
         self._params = {}
         for pname, pvalue in iteritems(params):
-            if '__' in pname:
+            if '__' in pname and _expand__to_dot:
                 pname = pname.replace('__', '.')
             self._setattr(pname, pvalue)
 
@@ -237,6 +255,8 @@ class DslBase(object):
                 # get the shortcut used to construct this type (query.Q, aggs.A, etc)
                 shortcut = self.__class__.get_dsl_type(pinfo['type'])
                 if pinfo.get('multi'):
+                    if not isinstance(value, (tuple, list)):
+                        value = (value, )
                     value = list(map(shortcut, value))
 
                 # dict(name -> DslBase), make sure we pickup all the objs
@@ -310,86 +330,13 @@ class DslBase(object):
     def _clone(self):
         return self._type_shortcut(self.to_dict())
 
-    def __add__(self, other):
-        # make sure we give queries that know how to combine themselves
-        # preference
-        if hasattr(other, '__radd__'):
-            return other.__radd__(self)
-        return self._bool(must=[self, other])
-
-    def __invert__(self):
-        return self._bool(must_not=[self])
-
-    def __or__(self, other):
-        # make sure we give queries that know how to combine themselves
-        # preference
-        if hasattr(other, '__ror__'):
-            return other.__ror__(self)
-        return self._bool(should=[self, other])
-
-    def __and__(self, other):
-        # make sure we give queries that know how to combine themselves
-        # preference
-        if hasattr(other, '__rand__'):
-            return other.__rand__(self)
-        return self._bool(must=[self, other])
-
-
-class BoolMixin(object):
-    """
-    Mixin containing all the operator overrides for Bool queries and filters.
-
-    Except for and where should behavior differs
-    """
-    def __add__(self, other):
-        q = self._clone()
-        if isinstance(other, self.__class__):
-            q.must += other.must
-            q.should += other.should
-            q.must_not += other.must_not
-        else:
-            q.must.append(other)
-        return q
-    __radd__ = __add__
-
-    def __or__(self, other):
-        if not (self.must or self.must_not):
-            # TODO: if only 1 in must or should, append the query instead of other
-            q = self._clone()
-            q.should.append(other)
-            return q
-
-        elif isinstance(other, self.__class__) and not (other.must or other.must_not):
-            # TODO: if only 1 in must or should, append the query instead of self
-            q = other._clone()
-            q.should.append(self)
-            return q
-
-        return self.__class__(should=[self, other])
-    __ror__ = __or__
-
-    def __invert__(self):
-        # special case for single negated query
-        if not (self.must or self.should) and len(self.must_not) == 1:
-            return self.must_not[0]._clone()
-
-        # bol without should, just flip must and must_not
-        elif not self.should:
-            q = self._clone()
-            q.must, q.must_not = q.must_not, q.must
-            return q
-
-        # TODO: should -> must_not.append(self.__class__(should=self.should)) ??
-        # queries with should just invert normally
-        return super(BoolMixin, self).__invert__()
-
 
 class ObjectBase(AttrDict):
     def __init__(self, **kwargs):
         m = self._doc_type.mapping
         for k in m:
             if k in kwargs and m[k]._coerce:
-                kwargs[k] = m[k].to_python(kwargs[k])
+                kwargs[k] = m[k].deserialize(kwargs[k])
         super(ObjectBase, self).__init__(kwargs)
 
     def __getattr__(self, name):
@@ -408,16 +355,18 @@ class ObjectBase(AttrDict):
 
     def __setattr__(self, name, value):
         if name in self._doc_type.mapping:
-            value = self._doc_type.mapping[name].to_python(value)
+            value = self._doc_type.mapping[name].deserialize(value)
         super(ObjectBase, self).__setattr__(name, value)
 
     def to_dict(self):
         out = {}
         for k, v in iteritems(self._d_):
-            if isinstance(v, (AttrList, list, tuple)):
-                v = [i.to_dict() if hasattr(i, 'to_dict') else i for i in v]
-            else:
-                v = v.to_dict() if hasattr(v, 'to_dict') else v
+            try:
+                f = self._doc_type.mapping[k]
+                if f._coerce:
+                    v = f.serialize(v)
+            except KeyError:
+                pass
 
             # don't serialize empty values
             # careful not to include numeric zeros
@@ -434,9 +383,12 @@ class ObjectBase(AttrDict):
             data = self._d_.get(name, None)
             try:
                 # save the cleaned value
-                self._d_[name] = field.clean(data)
+                data = field.clean(data)
             except ValidationException as e:
                 errors.setdefault(name, []).append(e)
+
+            if name in self._d_ or data not in ([], {}, None):
+                self._d_[name] = data
 
         if errors:
             raise ValidationException(errors)