+++ test_elasticsearch_dsl/test_document.py
@@ -1,4 +1,5 @@
 import pickle
+import codecs
 from hashlib import md5
 from datetime import datetime
 
@@ -30,7 +31,78 @@ class MyMultiSubDoc(MyDoc2, MySubDoc):
     pass
 
 class DocWithNested(document.DocType):
-    comments = field.Nested(properties={'title': field.String()})
+    comments = field.Nested(
+        properties={
+            'title': field.String(),
+            'tags': field.String(multi=True, index='not_analyzed')
+        }
+    )
+
+class SimpleCommit(document.DocType):
+    files = field.String(multi=True)
+
+    class Meta:
+        index = 'test-git'
+
+class Secret(str): pass
+
+class SecretField(field.CustomField):
+    builtin_type = 'string'
+
+    def _serialize(self, data):
+        return codecs.encode(data, 'rot_13')
+
+    def _deserialize(self, data):
+        if isinstance(data, Secret):
+            return data
+        return Secret(codecs.decode(data, 'rot_13'))
+
+class SecretDoc(document.DocType):
+    title = SecretField(index='no')
+
+class NestedSecret(document.DocType):
+    secrets = field.Nested(properties={'title': SecretField()})
+
+def test_custom_field():
+    s = SecretDoc(title=Secret('Hello'))
+
+    assert {'title': 'Uryyb'} == s.to_dict()
+    assert s.title == 'Hello'
+
+    s.title = 'Uryyb'
+    assert s.title == 'Hello'
+    assert isinstance(s.title, Secret)
+
+def test_custom_field_mapping():
+    assert {
+        'secret_doc': {
+            'properties': {
+                'title': {'index': 'no', 'type': 'string'}
+            }
+        }
+    } == SecretDoc._doc_type.mapping.to_dict()
+
+def test_custom_field_in_nested():
+    s = NestedSecret()
+    s.secrets.append({'title': Secret('Hello')})
+
+    assert {'secrets': [{'title': 'Uryyb'}]} == s.to_dict()
+    assert s.secrets[0].title == 'Hello'
+
+def test_multi_works_after_doc_has_been_saved():
+    c = SimpleCommit()
+    c.full_clean()
+    c.files.append('setup.py')
+
+    assert c.to_dict() == {'files': ['setup.py']}
+
+def test_multi_works_in_nested_after_doc_has_been_serialized():
+    # Issue #359
+    c = DocWithNested(comments=[{'title': 'First!'}])
+
+    assert [] == c.comments[0].tags
+    assert {'comments': [{'title': 'First!'}]} == c.to_dict()
+    assert [] == c.comments[0].tags
 
 def test_null_value_for_object():
     d = MyDoc(inner=None)
@@ -43,6 +115,7 @@ def test_inherited_doc_types_can_override_index():
             index = 'not-default-index'
 
     assert MyDocDifferentIndex._doc_type.index == 'not-default-index'
+    assert MyDocDifferentIndex().meta.index == 'not-default-index'
 
 def test_to_dict_with_meta():
     d = MySubDoc(title='hello')
@@ -213,7 +286,7 @@ def test_you_can_supply_own_mapping_instance():
         }
     } == MyD._doc_type.mapping.to_dict()
 
-def test_document_can_be_created_dynamicaly():
+def test_document_can_be_created_dynamically():
     n = datetime.now()
     md = MyDoc(title='hello')
     md.name = 'My Fancy Document!'
@@ -317,6 +390,34 @@ def test_search_with_custom_alias_and_index(mock_client):
     search_object = MyDoc.search(
       using="staging",
       index=["custom_index1", "custom_index2"])
-    
+
     assert search_object._using == "staging"
     assert search_object._index == ["custom_index1", "custom_index2"]
+
+def test_from_es_respects_underscored_non_meta_fields():
+    doc = {
+        "_index": "test-index",
+        "_type": "company",
+        "_id": "elasticsearch",
+        "_score": 12.0,
+
+        "fields": {
+            "hello": "world",
+            "_routing": "es",
+            "_tags": ["search"]
+
+        },
+
+        "_source": {
+            "city": "Amsterdam",
+            "name": "Elasticsearch",
+            "_tagline": "You know, for search"
+        }
+    }
+
+    class Company(document.DocType):
+        pass
+
+    c = Company.from_es(doc)
+
+    assert c.to_dict() == {'city': 'Amsterdam', 'hello': 'world', 'name': 'Elasticsearch', "_tags": ["search"], "_tagline": "You know, for search"}