+++ test_elasticsearch_dsl/test_utils.py
@@ -1,4 +1,19 @@
+import pickle
+
 from elasticsearch_dsl import utils, serializer
+
+def test_attrdict_pickle():
+    ad = utils.AttrDict({})
+
+    pickled_ad = pickle.dumps(ad)
+    assert ad == pickle.loads(pickled_ad)
+
+
+def test_attrlist_pickle():
+    al = utils.AttrList([])
+
+    pickled_al = pickle.dumps(al)
+    assert al == pickle.loads(pickled_al)
 
 def test_merge():
     a = {'a': {'b': 42, 'c': 47}}
@@ -35,3 +50,10 @@ def test_serializer_deals_with_Attr_versions():
     d = utils.AttrDict({'key': utils.AttrList([1, 2, 3])})
 
     assert serializer.serializer.dumps(d) == '{"key": [1, 2, 3]}'
+
+def test_serializer_deals_with_objects_with_to_dict():
+    class MyClass(object):
+        def to_dict(self):
+            return 42
+
+    assert serializer.serializer.dumps(MyClass()) == '42'